+++ next.config.ts
@@ -4,6 +4,7 @@ import createNextIntlPlugin from 'next-intl/plugin'
 const withNextIntl = createNextIntlPlugin('./src/i18n/request.ts')
 
 const nextConfig: NextConfig = {
+  allowedDevOrigins: ['192.168.4.22'],
   images: {
     formats: ['image/webp'],
     deviceSizes: [640, 750, 1080, 1200],